+++ nsflow/nde/flows/tests/test_realnvp.py
@@ -4,7 +4,7 @@ import torch
 import torchtestcase
 import unittest
 
-from nde.flows import realnvp
+from nsflow.nde.flows import realnvp
 
 
 class SimpleRealNVPTest(torchtestcase.TorchTestCase):
