+++ nsflow/nde/transforms/tests/test_qr.py
@@ -1,10 +1,10 @@
 import torch
 import unittest
 
-import utils
+from nsflow import utils
 
-from nde.transforms import qr
-from nde.transforms.transform_test import TransformTest
+from nsflow.nde.transforms import qr
+from nsflow.nde.transforms.tests.test_transform import TransformTest
 
 
 class QRLinearTest(TransformTest):
