+++ nsflow/nde/transforms/tests/test_svd.py
@@ -1,22 +1,23 @@
-"""Tests for the LU linear transforms."""
-
 import torch
 import unittest
 
-import utils
+from nsflow import utils
 
-from nde.transforms import lu
-from nde.transforms.transform_test import TransformTest
+from nsflow.nde.transforms import svd
+from nsflow.nde.transforms.tests.test_transform import TransformTest
 
 
-class LULinearTest(TransformTest):
+class SVDLinearTest(TransformTest):
 
     def setUp(self):
         self.features = 3
-        self.transform = lu.LULinear(features=self.features)
+        self.transform = svd.SVDLinear(features=self.features, num_householder=4)
+        self.transform.bias.data = torch.randn(self.features)  # Just so bias isn't zero.
 
-        lower, upper = self.transform._create_lower_upper()
-        self.weight = lower @ upper
+        diagonal = torch.diag(torch.exp(self.transform.log_diagonal))
+        orthogonal_1 = self.transform.orthogonal_1.matrix()
+        orthogonal_2 = self.transform.orthogonal_2.matrix()
+        self.weight = orthogonal_1 @ diagonal @ orthogonal_2
         self.weight_inverse = torch.inverse(self.weight)
         self.logabsdet = utils.logabsdet(self.weight)
